requirements.txt
deepgram-sdk
openai
python-dotenv
fastapi
uvicorn
websockets
jinja2